fix: Use f64 instead of f32
This increases the precision and shrinks the size of floating point
precision errors when calculating stuff like "10 / 3" or "1 / 3".

+++ calculator/src/ast.zig
@@ -7,7 +7,7 @@ pub const AstNode = struct {
         function_call: FunctionCall,
         nothing: u1,
         operand: struct {
-            number: f32,
+            number: f64,
             unit: ?[]const u8 = null,
         },
         variable_name: []const u8,
